LPD-5728 Convert from selector to just a function
