+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/plugins/page_content/components/PageContent.js
@@ -25,18 +25,16 @@ import {
 	useEditableProcessorUniqueId,
 	useSetEditableProcessorUniqueId,
 } from '../../../app/contexts/EditableProcessorContext';
-import {
-	useSelector,
-	useSelectorCallback,
-} from '../../../app/contexts/StoreContext';
+import {useSelector} from '../../../app/contexts/StoreContext';
 import selectCanUpdateEditables from '../../../app/selectors/selectCanUpdateEditables';
-import {selectPageContentDropdownItems} from '../../../app/selectors/selectPageContentDropdownItems';
 import getEditableId from '../../../app/utils/getEditableId';
 import getFirstControlsId from '../../../app/utils/getFirstControlsId';
 import getFragmentItem from '../../../app/utils/getFragmentItem';
+import {getPageContentDropdownItems} from '../../../app/utils/getPageContentDropdownItems';
 import ImageEditorModal from './ImageEditorModal';
 
 export default function PageContent({
+	actions,
 	classNameId,
 	classPK,
 	editableId,
@@ -67,40 +65,35 @@ export default function PageContent({
 		[editableId, editableProcessorUniqueId]
 	);
 
-	const dropdownItems = useSelectorCallback(
-		(state) => {
-			const pageContentDropdownItems = selectPageContentDropdownItems({
-				classNameId,
-				classPK,
-				externalReferenceCode,
-			})(state);
+	const dropdownItems = useMemo(() => {
+		const pageContentDropdownItems = getPageContentDropdownItems({
+			actions,
+		});
 
-			return pageContentDropdownItems?.map((item) => {
-				if (item.label === Liferay.Language.get('edit-image')) {
-					const {
-						editImageURL,
-						fileEntryId,
-						previewURL,
-						...editImageItem
-					} = item;
+		return pageContentDropdownItems?.map((item) => {
+			if (item.label === Liferay.Language.get('edit-image')) {
+				const {
+					editImageURL,
+					fileEntryId,
+					previewURL,
+					...editImageItem
+				} = item;
 
-					return {
-						...editImageItem,
-						onClick: () => {
-							setImageEditorParams({
-								editImageURL,
-								fileEntryId,
-								previewURL,
-							});
-						},
-					};
-				}
+				return {
+					...editImageItem,
+					onClick: () => {
+						setImageEditorParams({
+							editImageURL,
+							fileEntryId,
+							previewURL,
+						});
+					},
+				};
+			}
 
-				return item;
-			});
-		},
-		[classNameId, classPK, externalReferenceCode]
-	);
+			return item;
+		});
+	}, [actions]);
 
 	useEffect(() => {
 		if (editableProcessorUniqueId || !nextEditableProcessorUniqueId) {
